app/requirements.txt
flask
pandas
sqlalchemy
psycopg2-binary
requests